+++ crates/module-system/sov-test-utils/src/test_rollup.rs
@@ -257,6 +257,7 @@ impl<R: FullNodeBlueprint<Native> + Default + 'static> RollupBuilder<R> {
                         self.config.rollup_prover_config.clone(),
                         self.config.start_at_rollup_height,
                         self.config.stop_at_rollup_height,
+                        None,
                     )
                     .await?
             }
@@ -268,6 +269,7 @@ impl<R: FullNodeBlueprint<Native> + Default + 'static> RollupBuilder<R> {
                         self.config.rollup_prover_config.clone(),
                         self.config.start_at_rollup_height,
                         self.config.stop_at_rollup_height,
+                        None,
                     )
                     .await?
             }