+++ examples/demo-rollup/stf/src/runtime.rs
@@ -98,6 +98,9 @@ where
     #[cfg(feature = "native")]
     type GenesisInput = GenesisPaths;
 
+    #[cfg(feature = "native")]
+    type ModuleExecutionConfig = ();
+
     type Auth = sov_evm::EvmAuthenticator<S, Self>;
 
     #[cfg(feature = "native")]